@@ -1,52 +1,63 @@
+import { Component, OnInit, OnDestroy } from '@angular/core';
+import { Router, NavigationEnd, ActivatedRoute } from '@angular/router';
 import { TranslateService } from '@ngx-translate/core';
-import { Component, Output, TemplateRef, ViewChild, OnInit } from '@angular/core';
-import { BsModalService } from 'ngx-bootstrap/modal';
-import { BsModalRef } from 'ngx-bootstrap/modal/bs-modal-ref.service';
-import { FacebookService, InitParams } from 'ngx-facebook';
-import { Router, NavigationEnd } from '@angular/router';
+import { Subscription } from 'rxjs';
+import {
+  faPhone,
+  faEnvelope,
+  faCircleXmark,
+} from '@fortawesome/free-solid-svg-icons';
+import { CookieService } from 'ngx-cookie-service';
 
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
-  styleUrls: ['./app.component.css']
+  styleUrls: ['./app.component.sass'],
 })
-export class AppComponent implements OnInit {
-  @ViewChild('cookieLaw')
-  private cookieLawEl: any;
-
-  public dismiss(): void {
-    this.cookieLawEl.dismiss();
-  }
-
-  modalRef: BsModalRef;
+export class AppComponent implements OnInit, OnDestroy {
+  faPhone = faPhone;
+  faEnvelope = faEnvelope;
+  faCircleXmark = faCircleXmark;
+  cookieAllowed = false;
   items: any[];
 
-  constructor(private modalService: BsModalService, private fb: FacebookService, private router: Router) {
-    this.router.events.subscribe(event => {
+  private subscription: Subscription = new Subscription();
+
+  constructor(
+    private translate: TranslateService,
+    private activatedRoute: ActivatedRoute,
+    private router: Router,
+    private cookieService: CookieService
+  ) {
+    translate.setDefaultLang('it');
+    this.router.events.subscribe((event) => {
       if (event instanceof NavigationEnd) {
         (<any>window).ga('set', 'page', event.urlAfterRedirects);
         (<any>window).ga('send', 'pageview');
       }
     });
     this.items = Array(15).fill(0);
-    const initParams: InitParams = {
-      appId: '200610240719668',
-      xfbml: true,
-      version: 'v2.9'
-    };
-
-    fb.init(initParams);
   }
 
   ngOnInit() {
+    this.subscription = this.activatedRoute.queryParams.subscribe(
+      (param: any) => {
+        const locale = param['locale'];
+        if (locale !== undefined) {
+          this.translate.use(locale);
+        }
+      }
+    );
+
+    this.cookieAllowed = Boolean(this.cookieService.get('cantinaNegroCookie'));
   }
 
-  openModalCookie(template: TemplateRef<any>) {
-    this.cookieLawEl.dismiss();
-    this.modalRef = this.modalService.show(template);
+  ngOnDestroy() {
+    this.subscription.unsubscribe();
   }
 
-  openModal(template: TemplateRef<any>) {
-    this.modalRef = this.modalService.show(template);
+  allowCookies() {
+    this.cookieService.set('cantinaNegroCookie', 'true');
+    this.cookieAllowed = true;
   }
 }
